Working on adding assistant API.
This commit involves a significant modification in the architecture of the AI services. The AiService has been deleted and replaced by EmoDetectionServiceWithAssistant and AiAssistantService. These changes also include the addition of new properties in the `User` data model, and updates to `EmotionDetectionResult`, among others. New routes and database changes have been included as well.